#45 Moved the connection tests to a separate file
