+++ test/integration/databases_test.py
@@ -1,10 +1,9 @@
 import pytest
 
-from exasol.saas.client import openapi, PROMISING_STATES
+from exasol.saas.client import PROMISING_STATES
 from tenacity import RetryError
 from datetime import datetime, timedelta
 
-import pyexasol
 from exasol.saas.client.api_access import wait_for_delete_clearance, get_connection_params
 
 
@@ -51,35 +50,3 @@ def test_get_connection(api_access, database_name):
         connection = api_access.get_connection(db.id, clusters[0].id)
         assert connection.db_username is not None and \
             connection.port == 8563
-
-
-def test_get_connection_params_with_id(saas_host, saas_pat, saas_account_id,
-                                       operational_saas_database_id,
-                                       allow_connection):
-    """
-    This integration test checks that opening a pyexasol connection to a SaaS DB with
-    known id and executing a query works.
-    """
-    connection_params = get_connection_params(host=saas_host,
-                                              account_id=saas_account_id,
-                                              pat=saas_pat,
-                                              database_id=operational_saas_database_id)
-    with pyexasol.connect(**connection_params) as pyconn:
-        result = pyconn.execute('SELECT 1;').fetchall()
-        assert result == [(1,)]
-
-
-def test_get_connection_params_with_name(saas_host, saas_pat, saas_account_id,
-                                         operational_saas_database_id, database_name,
-                                         allow_connection):
-    """
-    This integration test checks that opening a pyexasol connection to a SaaS DB with
-    known name and executing a query works.
-    """
-    connection_params = get_connection_params(host=saas_host,
-                                              account_id=saas_account_id,
-                                              pat=saas_pat,
-                                              database_name=database_name)
-    with pyexasol.connect(**connection_params) as pyconn:
-        result = pyconn.execute('SELECT 1;').fetchall()
-        assert result == [(1,)]
